Add quit and preview modes

+++ src/index.ts
@@ -4,17 +4,14 @@ import {Storage} from './storage'
 
 export = class Reflect extends Command {
 
-  static description = 'describe the command here'
+  static description = 'Make target folder equivalent to source'
 
   static flags = {
-    // add --version flag to show CLI version
-    version: flags.version({char: 'v'}),
-    help: flags.help({char: 'h'}),
-    // flag with a value (-n, --name=VALUE)
-    name: flags.string({char: 'n', description: 'name to print'}),
-    // flag with no value (-f, --force)
-    force: flags.boolean({char: 'f'}),
-    ignorecase: flags.boolean({char: 'i'})
+       version: flags.version({char: 'v'}),
+          help: flags.help({char: 'h'}),
+    ignorecase: flags.boolean({char: 'i', description: 'not implemented yet'}),
+       preview: flags.boolean({char: 'p', description: 'do not change any file'}),
+         quiet: flags.boolean({char: 'q', description: 'no console output'})
   }
 
   static args = [
@@ -28,32 +25,30 @@ export = class Reflect extends Command {
   async run() {
     /* parse command line */
     const {args, flags} = this.parse(Reflect)
-
     /* create entity collector and storages */
     const source = new Storage(args.source);
     const target = new Storage(args.target);
-
     /* storages meet each other to terminate duplicate entities */
     source.mirror = target;
     target.mirror = source;
-
     /* walk source and collect file info */
     await Promise.all([
       walk(source),
       walk(target)
     ]);
-
     /* skip files which will be overwritten */
     target.skipByName(source);
-
     /* print results */
-    source.print(`source:`);
-    target.print(`target:`);
-
-    /* delete obsolete file from target */
-    target.delete();
-
-    /* copy new/changed files from source to target*/
-    source.copy(target);
+    if (!flags.quiet) {
+      source.print(`source:`);
+      target.print(`target:`);
+    }
+    /* ready to do the job */
+    if (!flags.preview) {
+      /* delete obsolete file from target */
+      target.delete();
+      /* copy new/changed files from source to target*/
+      source.copy(target);
+    }
   }
 }
